01_data_generation/pyproject.toml
[project]
name = "dbt-synthetic-data-generator"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "tenacity>=8.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "datasets>=2.0.0",
    "textual>=1.0.0",
    "dbt-duckdb>=1.8.0",
]

[tool.uv]
dev-dependencies = []
